Fix #303 : using TWeakObjectPtr for AALSBaseCharacter::DebugComponent; Code clean up as suggested by rider; Fix compilation fail with ue5-main

+++ Source/ALSV4_CPP/Private/Character/Animation/ALSCharacterAnimInstance.cpp
@@ -623,9 +623,9 @@ void UALSCharacterAnimInstance::UpdateMovementValues(float DeltaSeconds)
 
 	// Set the Relative Acceleration Amount and Interp the Lean Amount.
 	RelativeAccelerationAmount = CalculateRelativeAccelerationAmount();
-	LeanAmount.LR = FMath::FInterpTo(LeanAmount.LR, RelativeAccelerationAmount.Y, DeltaSeconds,
+	LeanAmount.LR = FMath::FInterpTo(LeanAmount.LR, static_cast<float>(RelativeAccelerationAmount.Y), DeltaSeconds,
 	                                 Config.GroundedLeanInterpSpeed);
-	LeanAmount.FB = FMath::FInterpTo(LeanAmount.FB, RelativeAccelerationAmount.X, DeltaSeconds,
+	LeanAmount.FB = FMath::FInterpTo(LeanAmount.FB, static_cast<float>(RelativeAccelerationAmount.X), DeltaSeconds,
 	                                 Config.GroundedLeanInterpSpeed);
 
 	// Set the Walk Run Blend
